+++ cmd/neofs-ir/internal/validate/validate_test.go
@@ -111,6 +111,15 @@ fschain:
 `,
 			wantErr: true,
 		},
+		{
+			name: "unknown field morph",
+			config: `
+morph:
+  dial_timeout: 1m
+  reconnections_number: 5
+`,
+			wantErr: true,
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {