+++ cmd/neofs-node/config/fschain/config_test.go
@@ -1,8 +1,6 @@
 package fschainconfig_test
 
 import (
-	"os"
-	"path/filepath"
 	"testing"
 	"time"
 
@@ -39,32 +37,5 @@ func TestFSChainSection(t *testing.T) {
 
 	t.Run("ENV", func(t *testing.T) {
 		configtest.ForEnvFileType(path, fileConfigTest)
-	})
-
-	t.Run("compatibility with morph section", func(t *testing.T) {
-		data := []byte(`
-morph:
-  dial_timeout: 30s
-  cache_ttl: 15s
-  reconnections_number: 6
-  reconnections_delay: 6s
-  endpoints:
-    - wss://rpc1.morph.fs.neo.org:40341/ws
-    - wss://rpc2.morph.fs.neo.org:40341/ws
-`)
-
-		morphPath := filepath.Join(t.TempDir(), "morph.yaml")
-		require.NoError(t, os.WriteFile(morphPath, data, 0o640))
-
-		var p config.Prm
-
-		os.Clearenv()
-
-		c := config.New(p,
-			config.WithConfigFile(morphPath),
-			config.WithValidate(false),
-		)
-
-		fileConfigTest(c)
 	})
 }